split up http_test and move assertion helpers into a new package
